Refactor SDK
