+++ .gitignore
@@ -1,0 +1,5 @@
+settings.xml
+.idea/
+out/
+px-java.iml
+target/
