+++ src/com/mercadopago/gradle.properties
@@ -1,0 +1,7 @@
+VERSION_NAME=1.0.0
+VERSION_CODE=1000
+GROUP=com.mercadopago
+
+POM_NAME=mercadopago
+POM_ARTIFACT_ID=sdk
+POM_PACKAGING=jar
