+++ src/com/mercadopago/resources/datastructures/preference/BackUrls.java
@@ -14,6 +14,12 @@ public class BackUrls {
     @Size(max=600) private String pending = null;
     @Size(max=600) private String failure = null;
 
+    public BackUrls(String successUrl, String pendingUrl, String failureUrl){
+        this.success=successUrl;
+        this.pending=pendingUrl;
+        this.failure=failureUrl;
+    }
+
 
     public String getSuccess() {
         return success;
