+++ src/com/mercadopago/resources/datastructures/preference/ExcludedPaymentType.java
@@ -12,6 +12,13 @@ public class ExcludedPaymentType {
 
     @Size(max=256) private String id = null;
 
+    public ExcludedPaymentType(String id) {
+        this.id=id;
+    }
+
+    public void ExcludedPaymentType(String id){
+        this.id=id;
+    }
 
     public ExcludedPaymentType setId(String id) {
         this.id = id;
