+++ src/test/mercadopago/core/MPCoreUtilsTest.java
@@ -1,6 +1,6 @@
 package test.mercadopago.core;
 
-import com.mercadopago.SDK;
+import com.mercadopago.MercadoPago;
 import com.mercadopago.core.MPCoreUtils;
 import com.mercadopago.exceptions.MPException;
 import org.apache.commons.io.IOUtils;
@@ -11,7 +11,7 @@ import java.io.InputStream;
 import static org.junit.Assert.*;
 
 /**
- * Mercado Pago SDK
+ * Mercado Pago MercadoPago
  * MPCoreUtils test class
  *
  * Created by Eduardo Paoletta on 11/17/16.
@@ -35,11 +35,11 @@ public class MPCoreUtilsTest {
     public void validateURLTest() {
         assertTrue(MPCoreUtils.validateUrl("https://www.google.com"));
         assertTrue(MPCoreUtils.validateUrl("https://mail.google.com/mail/u/0/#inbox"));
-        assertTrue(MPCoreUtils.validateUrl(SDK.getBaseUrl() + "/checkout/Preferences/4564"));
+        assertTrue(MPCoreUtils.validateUrl(MercadoPago.SDK.getBaseUrl() + "/checkout/Preferences/4564"));
 
         assertFalse(MPCoreUtils.validateUrl("djsfhsdkfhsdkfjhs"));
         assertFalse(MPCoreUtils.validateUrl("http://mail.google.com/mail/u/0/#inbox"));
-        assertFalse(MPCoreUtils.validateUrl(SDK.getBaseUrl() + "/checkout/Preferences/¿?"));
+        assertFalse(MPCoreUtils.validateUrl(MercadoPago.SDK.getBaseUrl() + "/checkout/Preferences/¿?"));
     }
 
 }
